+++ .gitignore
@@ -1,4 +1,5 @@
 .nopublish
+*.ini
 
 ### Windows ###
