+++ buildAndTest.ini
@@ -1,0 +1,3 @@
+gradlePropertiesPath=gradle.properties
+modName=TekCaysAddon
+instanceModPath=/home/pierre/.local/share/PrismLauncher/instances/TKCYA/.minecraft/mods/
